refactor: rename user display file
